refactored TemplateFactoryRSS to TemplateFactoryXML
